backend/script/requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv
pymysql
google-generativeai
tqdm
